chore: clean up agent/ai code
