+++ src/ipc/agent-tools/agent-tools-client.ts
@@ -1,8 +1,8 @@
 import {
-  OptimiseToolSelectionRequest,
-  OptimiseToolSizeRequest,
-  SearchApiEndpointsRequest,
-} from "./agent-tools-listeners";
+  OptimizeToolSelectionRequest,
+  OptimizeToolSizeRequest,
+} from "@/lib/mcp/tools";
+import { SearchApiEndpointsRequest } from "./agent-tools-listeners";
 
 export const listApis = async () => {
   return window.agentTools.listApis();
@@ -12,12 +12,12 @@ export const searchApiEndpoints = async (args: SearchApiEndpointsRequest) => {
   return window.agentTools.searchApiEndpoints(args);
 };
 
-export const optimiseToolSize = async (args: OptimiseToolSizeRequest) => {
+export const optimiseToolSize = async (args: OptimizeToolSizeRequest) => {
   return window.agentTools.optimiseToolSize(args);
 };
 
 export const optimiseToolSelection = async (
-  args: OptimiseToolSelectionRequest,
+  args: OptimizeToolSelectionRequest,
 ) => {
   return window.agentTools.optimiseToolSelection(args);
 };
