+++ src/lib/mcp/tools/index.ts
@@ -1,0 +1,5 @@
+// Central export for all tool-related functionality
+export * from "./types";
+export * from "./fetcher";
+export * from "./optimizer";
+export * from "./persistence";
