+++ src/lib/mcp/tools/mapper.ts
